requirements_production.txt
Django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.5
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
gunicorn==21.2.0
django-storages==1.14.2
boto3==1.34.0
whitenoise==6.6.0